+++ includes/wc-core-functions.php
@@ -1771,6 +1771,11 @@ function wc_uasort_comparison( $a, $b ) {
  * @return int
  */
 function wc_ascii_uasort_comparison( $a, $b ) {
+	// 'setlocale' is required for compatibility with PHP 8.
+	// Without it, 'iconv' will return '?'s instead of transliterated characters.
+	$prev_locale = setlocale( LC_CTYPE, 0 );
+	setlocale( LC_ALL, 'C.UTF-8' );
+
 	// phpcs:disable WordPress.PHP.NoSilencedErrors.Discouraged
 	if ( function_exists( 'iconv' ) && defined( 'ICONV_IMPL' ) && @strcasecmp( ICONV_IMPL, 'unknown' ) !== 0 ) {
 		$a = @iconv( 'UTF-8', 'ASCII//TRANSLIT//IGNORE', $a );
@@ -1778,6 +1783,7 @@ function wc_ascii_uasort_comparison( $a, $b ) {
 	}
 	// phpcs:enable WordPress.PHP.NoSilencedErrors.Discouraged
 
+	setlocale( LC_ALL, $prev_locale );
 	return strcmp( $a, $b );
 }