+++ src/Packages.php
@@ -68,6 +68,24 @@ class Packages {
 			}
 			call_user_func( array( $package_class, 'init' ) );
 		}
+
+		// Proxies "activated_plugin" hook for embedded packages listen on WC plugin activation
+		// https://github.com/woocommerce/woocommerce/issues/28697.
+		if ( is_admin() ) {
+			$activated_plugin = get_transient( 'woocommerce_activated_plugin' );
+			if ( $activated_plugin ) {
+				delete_transient( 'woocommerce_activated_plugin' );
+
+				/**
+				 * WooCommerce is activated hook.
+				 *
+				 * @since 5.0.0
+				 * @param bool $activated_plugin Activated plugin path,
+				 *                               generally woocommerce/woocommerce.php.
+				 */
+				do_action( 'woocommerce_activated_plugin', $activated_plugin );
+			}
+		}
 	}
 
 	/**