+++ tests/e2e/core-tests/specs/index.js
@@ -10,6 +10,7 @@ const runInitialStoreSettingsTest = require( './activate-and-setup/setup.test' )
 
 // Shopper tests
 const runCartPageTest = require( './shopper/front-end-cart.test' );
+const runCheckoutApplyCouponsTest = require( './shopper/front-end-checkout-coupons.test');
 const runCheckoutPageTest = require( './shopper/front-end-checkout.test' );
 const runMyAccountPageTest = require( './shopper/front-end-my-account.test' );
 const runSingleProductPageTest = require( './shopper/front-end-single-product.test' );
@@ -34,6 +35,7 @@ const runSetupOnboardingTests = () => {
 
 const runShopperTests = () => {
 	runCartPageTest();
+	runCheckoutApplyCouponsTest();
 	runCheckoutPageTest();
 	runMyAccountPageTest();
 	runSingleProductPageTest();
@@ -59,6 +61,7 @@ module.exports = {
 	runInitialStoreSettingsTest,
 	runSetupOnboardingTests,
 	runCartPageTest,
+	runCheckoutApplyCouponsTest,
 	runCheckoutPageTest,
 	runMyAccountPageTest,
 	runSingleProductPageTest,