+++ tests/e2e/core-tests/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Merchant Apply Coupon tests
 - Added new config variable for Simple Product price to `tests/e2e/env/config/default.json`. Defaults to 9.99
 
+- Shopper Checkout Apply Coupon
+
 ## Fixed
 
 - Flaky Create Product, Coupon, and Order tests